added getFile route

+++ src/Entity/File.php
@@ -65,12 +65,12 @@ class File
         return $this;
     }
 
-    public function getSize(): ?int
+    public function getSize(): ?float
     {
         return $this->size;
     }
 
-    public function setSize(?int $size): self
+    public function setSize(?float $size): self
     {
         $this->size = $size;
